Move DataSink/DataOutput to presto-common
