+++ presto-common/src/main/java/com/facebook/presto/common/io/DataSink.java
@@ -11,9 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.orc;
-
-import com.facebook.presto.orc.stream.DataOutput;
+package com.facebook.presto.common.io;
 
 import java.io.IOException;
 import java.util.List;
