+++ presto-hive/src/main/java/com/facebook/presto/hive/datasink/OutputStreamDataSinkFactory.java
@@ -13,8 +13,8 @@
  */
 package com.facebook.presto.hive.datasink;
 
-import com.facebook.presto.orc.DataSink;
-import com.facebook.presto.orc.OutputStreamDataSink;
+import com.facebook.presto.common.io.DataSink;
+import com.facebook.presto.common.io.OutputStreamDataSink;
 import com.facebook.presto.spi.ConnectorSession;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
