+++ presto-hive/src/main/java/com/facebook/presto/hive/pagefile/PageFileFooterOutput.java
@@ -13,8 +13,8 @@
  */
 package com.facebook.presto.hive.pagefile;
 
+import com.facebook.presto.common.io.DataOutput;
 import com.facebook.presto.hive.HiveCompressionCodec;
-import com.facebook.presto.orc.stream.DataOutput;
 import com.google.common.collect.ImmutableList;
 import io.airlift.slice.Slice;
 import io.airlift.slice.SliceOutput;
