+++ presto-hive/src/main/java/com/facebook/presto/hive/pagefile/PageFileWriter.java
@@ -14,9 +14,9 @@
 package com.facebook.presto.hive.pagefile;
 
 import com.facebook.presto.common.Page;
+import com.facebook.presto.common.io.DataSink;
 import com.facebook.presto.hive.HiveCompressionCodec;
 import com.facebook.presto.hive.HiveFileWriter;
-import com.facebook.presto.orc.DataSink;
 import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.spi.page.PagesSerde;
 import io.airlift.units.DataSize;
