+++ presto-orc/src/main/java/com/facebook/presto/orc/stream/StreamDataOutput.java
@@ -13,6 +13,7 @@
  */
 package com.facebook.presto.orc.stream;
 
+import com.facebook.presto.common.io.DataOutput;
 import com.facebook.presto.orc.metadata.Stream;
 import io.airlift.slice.Slice;
 import io.airlift.slice.SliceOutput;
