+++ presto-raptor/src/main/java/com/facebook/presto/raptor/filesystem/HdfsOrcDataEnvironment.java
@@ -13,11 +13,11 @@
  */
 package com.facebook.presto.raptor.filesystem;
 
+import com.facebook.presto.common.io.DataSink;
+import com.facebook.presto.common.io.OutputStreamDataSink;
 import com.facebook.presto.hive.HdfsContext;
-import com.facebook.presto.orc.DataSink;
 import com.facebook.presto.orc.OrcDataSource;
 import com.facebook.presto.orc.OrcDataSourceId;
-import com.facebook.presto.orc.OutputStreamDataSink;
 import com.facebook.presto.raptor.storage.OrcDataEnvironment;
 import com.facebook.presto.raptor.storage.ReaderAttributes;
 import com.facebook.presto.spi.PrestoException;
