+++ presto-raptor/src/main/java/com/facebook/presto/raptor/storage/OrcFileWriter.java
@@ -16,10 +16,10 @@ package com.facebook.presto.raptor.storage;
 import com.facebook.airlift.json.JsonCodec;
 import com.facebook.presto.common.NotSupportedException;
 import com.facebook.presto.common.Page;
+import com.facebook.presto.common.io.DataSink;
 import com.facebook.presto.common.type.Type;
 import com.facebook.presto.common.type.TypeManager;
 import com.facebook.presto.common.type.TypeSignature;
-import com.facebook.presto.orc.DataSink;
 import com.facebook.presto.orc.OrcWriter;
 import com.facebook.presto.orc.OrcWriterOptions;
 import com.facebook.presto.orc.OrcWriterStats;
